refactor: CureSilver 사전신청 API 수정
