--- src/main/java/site/cleanfree/be_main/curesilver/CureSilverRegisterRequestVo.java
@@ -1,9 +1,0 @@
-package site.cleanfree.be_main.curesilver;
-
-import lombok.Getter;
-
-@Getter
-public class CureSilverRegisterRequestVo {
-
-    private String ip;
-}
