+++ src/main/java/site/cleanfree/be_main/curesilver/domain/CureSilverAccess.java
@@ -1,4 +1,4 @@
-package site.cleanfree.be_main.curesilver;
+package site.cleanfree.be_main.curesilver.domain;
 
 import jakarta.persistence.Id;
 import lombok.Builder;
@@ -8,18 +8,19 @@ import org.springframework.data.mongodb.core.mapping.Document;
 import site.cleanfree.be_main.common.MongoBaseTimeEntity;
 
 @Getter
-@Document(collection = "cure_silver")
-public class CureSilver extends MongoBaseTimeEntity {
+@Document(collection = "cure_silver_access")
+public class CureSilverAccess extends MongoBaseTimeEntity {
 
     @Id
     private String id;
-    @Indexed(unique = true)
+    @Indexed
     private String ip;
+    private int count;
 
     @Builder
-    public CureSilver(String id, String ip) {
+    public CureSilverAccess(String id, String ip, int count) {
         this.id = id;
         this.ip = ip;
+        this.count = count;
     }
-
 }
