+++ src/main/java/site/cleanfree/be_main/curesilver/infrastructure/CureSilverRepository.java
@@ -1,9 +1,9 @@
-package site.cleanfree.be_main.curesilver;
+package site.cleanfree.be_main.curesilver.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import site.cleanfree.be_main.curesilver.domain.CureSilver;
 
 public interface CureSilverRepository extends MongoRepository<CureSilver, String> {
 
-    Optional<CureSilver> findCureSilverByIp(String ip);
 }
