+++ src/main/java/site/cleanfree/be_main/curesilver/presentation/CureSilverController.java
@@ -1,4 +1,4 @@
-package site.cleanfree.be_main.curesilver;
+package site.cleanfree.be_main.curesilver.presentation;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -11,6 +11,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.cleanfree.be_main.common.BaseResponse;
 import site.cleanfree.be_main.common.ClientIpAccessor;
+import site.cleanfree.be_main.curesilver.vo.CureSilverRegisterRequestVo;
+import site.cleanfree.be_main.curesilver.application.CureSilverService;
 
 @RestController
 @RequestMapping("/api/v1/curesilver")
@@ -23,10 +25,8 @@ public class CureSilverController {
     @PostMapping("/register")
     @Operation(summary = "사전 신청 API", description = "ip 주소를 기준으로 사전 신청을 저장합니다.")
     public ResponseEntity<BaseResponse<Object>> register(
-        HttpServletRequest request,
         @RequestBody CureSilverRegisterRequestVo cureSilverRegisterRequestVo
     ) {
-        String clientIp = ClientIpAccessor.getIp(request);
-        return ResponseEntity.ok(cureSilverService.register(clientIp));
+        return ResponseEntity.ok(cureSilverService.register(cureSilverRegisterRequestVo));
     }
 }
