+++ src/main/java/site/cleanfree/be_main/curesilver/vo/CureSilverRegisterRequestVo.java
@@ -1,0 +1,5 @@
+package site.cleanfree.be_main.curesilver.vo;
+
+public record CureSilverRegisterRequestVo(String name, String phoneNumber) {
+
+}
